# server/swagger.yaml
openapi: 3.0.3
info:
  title: Bike Rental API
  version: "1.0.0"
  description: |-
    Simple bike rental backend (Express + MySQL using mysql2) — OpenAPI specification generated from the server source.
servers:
  - url: http://localhost:4000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: alice
        role:
          type: string
          example: user
...
    Station:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        lat:
          type: number
          format: float
        lng:
          type: number
          format: float
        capacity:
          type: integer
        available:
          type: integer
        open:
          type: boolean
...
    Report:
      type: object
      properties:
        id:
          type: integer
        description:
          type: string
        status:
          type: string
          example: open
        StationId:
          type: integer
        UserId:
          type: integer
    Rental:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
          example: active
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
        UserId:
          type: integer
        fromStationId:
          type: integer
        toStationId:
          type: integer
paths:
...
  /stations/{id}:
    get:
      summary: Get a station by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: station
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Station"
        "404":
          description: Not found
  /stations/{id}/toggle:
    put:
      summary: Toggle open/closed for a station
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: station updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Station"
        "404": { description: Not found }
...
  /reports:
    get:
      summary: List all reports
      responses:
        "200":
          description: list of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      summary: Create a report (authenticated)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                stationId: { type: integer }
                description: { type: string }
              required: [stationId, description]
      responses:
        "200":
          description: created report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /reports/me:
    get:
      summary: Get current user's reports
      security:
        - bearerAuth: []
      responses:
        "200":
          description: user reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
  /reports/{id}:
    put:
      summary: Update a report (owner)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                description: { type: string }
      responses:
        "200": { description: updated report }
        "403": { description: Forbidden }
        "404": { description: Not found }
    delete:
      summary: Delete a report (owner)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200": { description: success }
  /profile:
    get:
      summary: Get current user's profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      summary: Update current user's profile
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        "200": { description: updated user basic info }
...
  /rentals/rent:
    post:
      summary: Rent a bike from a station (authenticated)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                stationId: { type: integer }
              required: [stationId]
      responses:
        "200":
          description: rental started
          content:
            application/json:
              schema:
                type: object
                properties:
                  rentalId: { type: integer }
                  status: { type: string }
  /rentals/return:
    post:
      summary: Return a bike to a station (authenticated)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                rentalId: { type: integer }
                stationId: { type: integer }
              required: [rentalId, stationId]
      responses:
        "200":
          description: rental returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  rentalId: { type: integer }
                  status: { type: string }
  /rentals/me:
    get:
      summary: List current user's rentals
      security:
        - bearerAuth: []
      responses:
        "200":
          description: list of rentals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rental"
...
  /admin/reports:
    get:
      summary: Admin - list all reports
      security:
        - bearerAuth: []
      responses:
        "200":
          description: list of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
  /admin/reports/{id}:
    put:
      summary: Admin - update report status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string }
      responses:
        "200": { description: updated report }
    delete:
      summary: Admin - delete report
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200": { description: success }
...
  /admin/stations:
    get:
      summary: Admin - list stations
      security:
        - bearerAuth: []
      responses:
        "200":
          description: list stations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Station"
    post:
      summary: Admin - create a station
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                lat: { type: number }
                lng: { type: number }
                capacity: { type: integer }
                available: { type: integer }
                open: { type: boolean }
      responses:
        "200": { description: created station }
  /admin/stations/{id}:
    put:
      summary: Admin - update a station
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200": { description: updated station }
    delete:
      summary: Admin - delete a station
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200": { description: success }
...